merge

+++ tests/src/test/java/com/orientechnologies/orient/test/database/auto/DbImportExportTest.java
@@ -45,12 +45,13 @@ public class DbImportExportTest extends DocumentDBBaseTest implements OCommandOu
   public DbImportExportTest(@Optional String url, String testPath) {
     super(url);
     this.testPath = testPath;
-    this.exportFilePath = System.getProperty("exportFilePath", EXPORT_FILE_PATH);
+
+    exportFilePath = System.getProperty("exportFilePath", EXPORT_FILE_PATH);
   }
 
   @Test
   public void testDbExport() throws IOException {
-    final ODatabaseDocumentTx database = new ODatabaseDocumentTx(url);
+    ODatabaseDocumentTx database = new ODatabaseDocumentTx(url);
     database.open("admin", "admin");
 
     // ADD A CUSTOM TO THE CLASS
@@ -58,35 +59,30 @@ public class DbImportExportTest extends DocumentDBBaseTest implements OCommandOu
         .command(new OCommandSQL("alter class V custom onBeforeCreate=onBeforeCreateItem"))
         .execute();
 
-    final ODatabaseExport export =
-        new ODatabaseExport(database, testPath + "/" + exportFilePath, this);
+    ODatabaseExport export = new ODatabaseExport(database, testPath + "/" + exportFilePath, this);
     export.exportDatabase();
     export.close();
+
     database.close();
   }
 
   @Test(dependsOnMethods = "testDbExport")
   public void testDbImport() throws IOException {
     final File importDir = new File(testPath + "/" + NEW_DB_PATH);
-    if (importDir.exists()) {
-      for (final File f : importDir.listFiles()) {
-        f.delete();
-      }
-    } else {
-      importDir.mkdir();
-    }
+    if (importDir.exists()) for (File f : importDir.listFiles()) f.delete();
+    else importDir.mkdir();
 
-    final ODatabaseDocumentTx database =
+    ODatabaseDocumentTx database =
         new ODatabaseDocumentTx(getStorageType() + ":" + testPath + "/" + NEW_DB_URL);
     database.create();
 
-    final ODatabaseImport dbImport =
-        new ODatabaseImport(database, testPath + "/" + exportFilePath, this);
+    ODatabaseImport dbImport = new ODatabaseImport(database, testPath + "/" + exportFilePath, this);
 
     // UNREGISTER ALL THE HOOKS
-    for (final ORecordHook hook : new ArrayList<>(database.getHooks().keySet())) {
+    for (ORecordHook hook : new ArrayList<ORecordHook>(database.getHooks().keySet())) {
       database.unregisterHook(hook);
     }
+
     dbImport.setPreserveRids(true);
     dbImport.setDeleteRIDMapping(false);
     dbImport.importDatabase();
@@ -98,13 +94,13 @@ public class DbImportExportTest extends DocumentDBBaseTest implements OCommandOu
   @Test(dependsOnMethods = "testDbImport")
   public void testCompareDatabases() throws IOException {
     if ("remote".equals(getStorageType()) || url.startsWith("remote:")) {
-      final String env = getTestEnv();
-      if (env == null || env.equals("dev")) {
-        return;
-      }
+      String env = getTestEnv();
+      if (env == null || env.equals("dev")) return;
+
       // EXECUTES ONLY IF NOT REMOTE ON CI/RELEASE TEST ENV
     }
-    final String urlPrefix = getStorageType() + ":";
+
+    String urlPrefix = getStorageType() + ":";
 
     final ODatabaseCompare databaseCompare =
         new ODatabaseCompare(
@@ -122,12 +118,10 @@ public class DbImportExportTest extends DocumentDBBaseTest implements OCommandOu
   @Override
   @Test(enabled = false)
   public void onMessage(final String iText) {
-    if (iText != null && iText.contains("ERR")) {
+    if (iText != null && iText.contains("ERR"))
       // ACTIVATE DUMP MODE
       dumpMode = true;
-    }
-    if (dumpMode) {
-      OLogManager.instance().error(this, iText, null);
-    }
+
+    if (dumpMode) OLogManager.instance().error(this, iText, null);
   }
 }
